Change out template_dll solution files so that it generates the DLL with the correct name and in the correct location
