--- external/source/exploits/CVE-2020-0787/template_dll/WindowsCoreDeviceInfo.def
@@ -1,4 +1,0 @@
-LIBRARY WindowsCoreDeviceInfo 
-
-EXPORTS
-QueryDeviceInformation PRIVATE
