--- external/source/exploits/CVE-2020-0787/template_dll/WindowsCoreDeviceInfo.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WindowsCoreDeviceInfo", "WindowsCoreDeviceInfo.vcxproj", "{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Debug|x64.ActiveCfg = Debug|x64
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Debug|x64.Build.0 = Debug|x64
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Debug|x86.ActiveCfg = Debug|Win32
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Debug|x86.Build.0 = Debug|Win32
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Release|x64.ActiveCfg = Release|x64
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Release|x64.Build.0 = Release|x64
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Release|x86.ActiveCfg = Release|Win32
-		{B27BF6FA-A0D4-4891-BBB1-12CD7019AE3D}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {02E2890C-192C-44D3-B684-D617515A3A8D}
-	EndGlobalSection
-EndGlobal
