+++ external/source/exploits/CVE-2020-0787/template_dll/template_dll.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "template_dll", "template_dll\template_dll.vcxproj", "{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Debug|x64.ActiveCfg = Debug|x64
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Debug|x64.Build.0 = Debug|x64
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Debug|x86.ActiveCfg = Debug|Win32
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Debug|x86.Build.0 = Debug|Win32
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Release|x64.ActiveCfg = Release|x64
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Release|x64.Build.0 = Release|x64
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Release|x86.ActiveCfg = Release|Win32
+		{5C2D5EB2-6203-4C40-9DDC-AE07C13CCE1A}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F52BBD00-5ABD-4D96-AD38-06AB78D7CD23}
+	EndGlobalSection
+EndGlobal
