+++ external/source/exploits/CVE-2020-0787/template_dll/template_dll/framework.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
+// Windows Header Files
+#include <windows.h>
