+++ external/source/exploits/CVE-2020-0787/template_dll/template_dll/template_dll.def
@@ -1,0 +1,4 @@
+LIBRARY template_dll
+
+EXPORTS
+QueryDeviceInformation PRIVATE
